requirements.txt
asgiref==3.6.0
Django==4.1.5
django-s3-sqlite==0.0.5
django-taggit==3.1.0
mysqlclient==2.1.1
Pillow==9.4.0
sqlparse==0.4.3
